Fix race
